requirements.txt
fastapi==0.105.0
uvicorn==0.24.0  
motor==3.6.0
pydantic==2.4.2
google-cloud-storage==2.12.0
python-multipart==0.0.6
aiofiles==23.2.1